+++ code/agent/monitor/data/orchestrator_data.py
@@ -33,6 +33,7 @@ class ClusterStatusData(BaseDataStructure):
     cluster_managers: Union[List, None]
     cluster_join_address: Union[str, None]
     cluster_nodes: Union[List, None]
+    labels: Union[List, None]
 
 
 class DeploymentData(BaseDataStructure):